feat: start creating useForm hook
